исправлены теперь все недочеты при деактивации аккаунта.

+++ public/js/userLk.js
@@ -481,9 +481,13 @@ forgetMeBttn.onclick = function(event){
 		if(pswdInput.value == userApi.curUser.password){
             console.log("Деактивация");
             userApi.deactivateAccount(function (success) {
-                socket.emit('leave', { username: userApi.curUser.login });
-                localStorage.clear();
-                window.location = "/";
+                if (success) {
+                    socket.emit('leave', { username: userApi.curUser.login });
+                    localStorage.clear();
+                    window.location = "/";
+                } else {
+                    alert("Не удалось деактивировать аккаунт!");
+                }
             });
 		}
 		else{
